backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP client for NVIDIA API
httpx==0.27.0
aiohttp==3.9.5

# Export utilities
python-docx==1.1.2
WeasyPrint==62.3
markdown2==2.4.13

# Utilities
python-dotenv==1.0.1
tenacity==8.3.0
structlog==24.2.0
